still stuck

+++ src/Util/MyMathUtils.java
@@ -127,26 +127,26 @@ public class MyMathUtils {
         return ret;
     }
 
-    public static void applyGradientDescentToBiases(RealVector[] biases, RealVector[] biasNablas, int learningRate, int batchSize)
+    public static void applyGradientDescentToBiases(RealVector[] biases, RealVector[] biasNablas, double learningRate, double batchSize)
     {
         for (int layerIndex = 0; layerIndex < biases.length; layerIndex++) {
             RealVector biasesAtLayer = biases[layerIndex];
             for (int biasIndex = 0; biasIndex < biasesAtLayer.getDimension(); biasIndex++) {
                 double bias = biasesAtLayer.getEntry(biasIndex);
-                double newBias =  bias - ((double)learningRate / (double)batchSize * biasNablas[layerIndex].getEntry(biasIndex));
+                double newBias =  bias - (learningRate / batchSize * biasNablas[layerIndex].getEntry(biasIndex));
                 biasesAtLayer.setEntry(biasIndex, newBias);
             }
         }
     }
 
-    public static void applyGradientDescentToWeights(RealMatrix[] weights, RealMatrix[] weightNablas, int learningRate, int batchSize)
+    public static void applyGradientDescentToWeights(RealMatrix[] weights, RealMatrix[] weightNablas, double learningRate, double batchSize)
     {
         for (int layerIndex = 0; layerIndex < weightNablas.length; layerIndex++) {
             RealMatrix weightsAtLayer = weights[layerIndex];
             for (int i = 0; i < weightsAtLayer.getRowDimension(); i++) {
                 for (int j = 0; j < weightsAtLayer.getColumnDimension(); j++) {
                     double weight = weightsAtLayer.getEntry(i, j);
-                    double newWeight = weight - ((double)learningRate / (double)batchSize * weightNablas[layerIndex].getEntry(i, j));
+                    double newWeight = weight - (learningRate / batchSize * weightNablas[layerIndex].getEntry(i, j));
                     weightsAtLayer.setEntry(i, j, newWeight);
                 }
             }
